Task 16, first 98 fibonacci numbers

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,13 +9,13 @@
 int main(void)
 {
 	int i = 2;
-	long long int fib_1 = 0;
-	long long int fib_2 = 1;
-	long long int fib_n = fib_1 + fib_2;
+	unsigned long int fib_1 = 0;
+	unsigned long int fib_2 = 1;
+	unsigned long int fib_n = fib_1 + fib_2;
 
 	while (i <= 98)
 	{
-		printf("%lld, ", fib_n);
+		printf("%lu, ", fib_n);
 		fib_1 = fib_2;
 		fib_2 = fib_n;
 		fib_n = fib_1 + fib_2;
@@ -23,7 +23,7 @@ int main(void)
 		i++;
 	}
 	fib_n = fib_1 + fib_2;
-	printf("%lld\n", fib_n);
+	printf("%lu\n", fib_n);
 
 	return (0);
 }
